src/requirement.txt
aiofiles==23.2.1
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
altair==5.4.0
annotated-types==0.7.0
ansi2html==1.9.1
anyio==4.4.0
appnope==0.1.3
asgiref==3.8.1
asttokens==2.2.1
async-timeout==4.0.3
attrs==24.2.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.2.0
build==1.2.1
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.1.3
coloredlogs==15.0.1
comm==0.1.2
contourpy==1.2.1
cycler==0.12.1
dash==2.14.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.7
db-dtypes==1.4.3
debugpy==1.6.4
decorator==5.1.1
Deprecated==1.2.14
entrypoints==0.4
exceptiongroup==1.2.2
executing==1.2.0
fastapi==0.112.0
ffmpy==0.4.0
filelock==3.15.4
Flask==2.1.2
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.2.2
google-cloud-bigquery==3.10.0
google-cloud-core==2.3.2
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.57.0
gpt4all==2.8.0
gradio==3.36.1
gradio_client==1.3.0
greenlet==3.0.3
grpcio==1.51.3
grpcio-status==1.51.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.21.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib_resources==6.4.0
ipykernel==6.19.2
ipython==8.7.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==7.4.8
jupyter_core==5.1.0
kiwisolver==1.4.5
kubernetes==30.1.0
langchain==0.2.12
langchain-chroma==0.1.2
langchain-community==0.2.11
langchain-core==0.2.29
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.2
langchainhub==0.1.20
langsmith==0.1.98
linkify-it-py==2.0.3
logging==0.4.9.6
markdown-it-py==2.2.0
MarkupSafe==2.1.1
marshmallow==3.21.3
matplotlib==3.9.1.post1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
narwhals==1.4.2
nest-asyncio==1.5.6
networkx==3.2.1
nltk==3.9
numpy==1.24.4
oauthlib==3.2.2
ollama==0.3.1
onnxruntime==1.18.1
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.10.7
overrides==7.7.0
packaging==24.2
pandas==2.0.3
pandas-gbq==0.29.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==2.6.0
plotly==5.18.0
posthog==3.5.0
prompt-toolkit==3.0.36
proto-plus==1.26.1
protobuf==4.21.12
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==21.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==2.8.2
pydantic_core==2.20.1
pydata-google-auth==1.9.1
pydub==0.25.1
Pygments==2.13.0
pyparsing==3.0.9
PyPika==0.48.9
pyproject_hooks==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2023.3
PyYAML==6.0.2
pyzmq==24.0.1
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
ruff==0.5.7
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.10.1
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
stack-data==0.6.2
starlette==0.37.2
sympy==1.13.2
tenacity==8.2.3
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.2
tqdm==4.66.5
traitlets==5.7.1
transformers==4.44.0
typer==0.12.3
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2024.1
uc-micro-py==1.0.3
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.5
uvloop==0.19.0
watchfiles==0.23.0
wcwidth==0.2.5
websocket-client==1.8.0
websockets==12.0
Werkzeug==2.1.2
wrapt==1.16.0
yarl==1.9.4
zipp==3.15.0
